fix: attach wp-codebox.zip as release asset
release:package only declared the npm CLI tarball in its stdout manifest,
so the release pipeline never uploaded packages/wordpress-plugin/dist/wp-codebox.zip
even though homeboy.json declares it as build_artifact. Default deploys
(release_asset source) then 404'd on the missing zip.

Build the plugin zip during release:package and surface it in the manifest
so it is uploaded as a GitHub Release asset named exactly wp-codebox.zip.
Extract the zip-assembly into a shared lib to avoid the release:package
recursion that build-wordpress-plugin-zip.ts would otherwise introduce.

Closes #1592
